No longer split kernel builder class
