+++ firedrake/slate/slac/kernel_builder.py
@@ -1,10 +1,9 @@
-from abc import ABCMeta, abstractproperty
-
 from coffee import base as ast
 
 from collections import OrderedDict, Counter, namedtuple
 
-from firedrake.slate.slac.utils import topological_sort, traverse_dags
+from firedrake.slate.slac.utils import (topological_sort, traverse_dags,
+                                        eigen_tensor, Transformer)
 from firedrake.utils import cached_property
 
 from functools import reduce
@@ -32,11 +31,45 @@ Context information for creating coefficient temporaries.
 """
 
 
-class KernelBuilderBase(object, metaclass=ABCMeta):
-    """A base helper class for constructing Slate kernels."""
+class LocalKernelBuilder(object):
+    """The primary helper class for constructing cell-local linear
+    algebra kernels from Slate expressions.
+
+    This class provides access to all temporaries and subkernels associated
+    with a Slate expression. If the Slate expression contains nodes that
+    require operations on already assembled data (such as the action of a
+    slate tensor on a `ufl.Coefficient`), this class provides access to the
+    expression which needs special handling.
+
+    Instructions for assembling the full kernel AST of a Slate expression is
+    provided by the method `construct_ast`.
+    """
+
+    # Relevant symbols/information needed for kernel construction
+    # defined below
+    coord_sym = ast.Symbol("coords")
+    cell_orientations_sym = ast.Symbol("cell_orientations")
+    cell_facet_sym = ast.Symbol("cell_facets")
+    it_sym = ast.Symbol("i0")
+    mesh_layer_sym = ast.Symbol("layer")
+
+    # Supported integral types
+    supported_integral_types = [
+        "cell",
+        "interior_facet",
+        "exterior_facet",
+        # The "interior_facet_horiz" measure is separated into two parts:
+        # "top" and "bottom"
+        "interior_facet_horiz_top",
+        "interior_facet_horiz_bottom",
+        "interior_facet_vert",
+        "exterior_facet_top",
+        "exterior_facet_bottom",
+        "exterior_facet_vert"
+    ]
 
     def __init__(self, expression, tsfc_parameters=None):
-        """Constructor for the KernelBuilderBase class.
+        """Constructor for the LocalKernelBuilder class.
 
         :arg expression: a :class:`TensorBase` object.
         :arg tsfc_parameters: an optional `dict` of parameters to provide to
@@ -94,6 +127,75 @@ class KernelBuilderBase(object, metaclass=ABCMeta):
                           and not isinstance(tensor, (slate.Tensor,
                                                       slate.Negative))]
         self.action_coefficients = action_coeffs
+        self._setup()
+
+    def _setup(self):
+        """A setup method to initialize all the local assembly
+        kernels generated by TSFC and creates templated function calls
+        conforming to the Eigen-C++ template library standard.
+        This function also collects any information regarding orientations
+        and extra include directories.
+        """
+        transformer = Transformer()
+        include_dirs = []
+        templated_subkernels = []
+        assembly_calls = OrderedDict([(it, [])
+                                      for it in self.supported_integral_types])
+        coords = None
+        oriented = False
+        for cxt_kernel in self.context_kernels:
+            local_coefficients = cxt_kernel.coefficients
+            it_type = cxt_kernel.original_integral_type
+            exp = cxt_kernel.tensor
+
+            if it_type not in self.supported_integral_types:
+                raise ValueError("Integral type '%s' not recognized" % it_type)
+
+            # Explicit checking of coordinates
+            coordinates = cxt_kernel.tensor.ufl_domain().coordinates
+            if coords is not None:
+                assert coordinates == coords, "Mismatching coordinates!"
+            else:
+                coords = coordinates
+
+            for split_kernel in cxt_kernel.tsfc_kernels:
+                indices = split_kernel.indices
+                kinfo = split_kernel.kinfo
+
+                # TODO: Implement subdomains for Slate tensors
+                if kinfo.subdomain_id != "otherwise":
+                    raise NotImplementedError("Subdomains not implemented.")
+
+                args = [c for i in kinfo.coefficient_map
+                        for c in self.coefficient(local_coefficients[i])]
+
+                if kinfo.oriented:
+                    args.insert(0, self.cell_orientations_sym)
+
+                if kinfo.integral_type in ["interior_facet",
+                                           "exterior_facet",
+                                           "interior_facet_vert",
+                                           "exterior_facet_vert"]:
+                    args.append(ast.FlatBlock("&%s" % self.it_sym))
+
+                # Assembly calls within the macro kernel
+                tensor = eigen_tensor(exp, self.temps[exp], indices)
+                call = ast.FunCall(kinfo.kernel.name,
+                                   tensor,
+                                   self.coord_sym,
+                                   *args)
+                assembly_calls[it_type].append(call)
+
+                # Subkernels for local assembly (Eigen templated functions)
+                kast = transformer.visit(kinfo.kernel._ast)
+                templated_subkernels.append(kast)
+                include_dirs.extend(kinfo.kernel._include_dirs)
+                oriented = oriented or kinfo.oriented
+
+        self.assembly_calls = assembly_calls
+        self.templated_subkernels = templated_subkernels
+        self.include_dirs = list(set(include_dirs))
+        self.oriented = oriented
 
     @cached_property
     def coefficient_map(self):
@@ -136,9 +238,33 @@ class KernelBuilderBase(object, metaclass=ABCMeta):
                        for cxt_k in cxt_tuple]
         return cxt_kernels
 
-    @abstractproperty
+    @property
     def integral_type(self):
-        """Returns the integral type associated with a Slate kernel. This
-        is used to determine how the Slate kernel should be iterated over
-        the mesh.
+        """Returns the integral type associated with a Slate kernel."""
+        return "cell"
+
+    @cached_property
+    def needs_cell_facets(self):
+        """Searches for any embedded forms (by inspecting the ContextKernels)
+        which require looping over cell facets. If any are found, this function
+        returns `True` and `False` otherwise.
         """
+        cell_facet_types = ["interior_facet",
+                            "exterior_facet",
+                            "interior_facet_vert",
+                            "exterior_facet_vert"]
+        return any(cxt_k.original_integral_type in cell_facet_types
+                   for cxt_k in self.context_kernels)
+
+    @cached_property
+    def needs_mesh_layers(self):
+        """Searches for any embedded forms (by inspecting the ContextKernels)
+        which require mesh level information (extrusion measures). If any are
+        found, this function returns `True` and `False` otherwise.
+        """
+        mesh_layer_types = ["interior_facet_horiz_top",
+                            "interior_facet_horiz_bottom",
+                            "exterior_facet_bottom",
+                            "exterior_facet_top"]
+        return any(cxt_k.original_integral_type in mesh_layer_types
+                   for cxt_k in self.context_kernels)
